test(agents): split scheduled projection coverage
